# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for E-Document Management System",
        "title": "E-Document API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@edocument.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:5001",
    "basePath": "/api",
    "paths": {
...
        "/v1/incoming-docs": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all incoming documents (for admin/staff)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "incoming-docs"
                ],
                "summary": "Get all incoming documents",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/e-document-backend_internal_util.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/e-document-backend_internal_domain.IncomingDocResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/e-document-backend_internal_util.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/e-document-backend_internal_util.Response"
                        }
                    }
                }
            }
        },
        "/v1/incoming-docs/department/{deptId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get incoming documents assigned to a specific department (UUID)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "incoming-docs"
                ],
                "summary": "Get incoming documents by department",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Department UUID",
                        "name": "deptId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/e-document-backend_internal_util.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/e-document-backend_internal_domain.IncomingDocResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/e-document-backend_internal_util.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/e-document-backend_internal_util.Response"
                        }
                    }
                }
            }
        },
        "/v1/incoming-docs/receiver/{receiverId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get incoming documents assigned to a specific receiver (user UUID)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "incoming-docs"
                ],
                "summary": "Get incoming documents by receiver",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Receiver User UUID",
                        "name": "receiverId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/e-document-backend_internal_util.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/e-document-backend_internal_domain.IncomingDocResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/e-document-backend_internal_util.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/e-document-backend_internal_util.Response"
                        }
                    }
                }
            }
        },
        "/v1/incoming-docs/status/{status}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get incoming documents filtered by status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "incoming-docs"
                ],
                "summary": "Get incoming documents by status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Status (pending, received, approved, rejected)",
                        "name": "status",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/e-document-backend_internal_util.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/e-document-backend_internal_domain.IncomingDocResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/e-document-backend_internal_util.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/e-document-backend_internal_util.Response"
                        }
                    }
                }
            }
        },
        "/v1/incoming-docs/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get incoming document details by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "incoming-docs"
                ],
                "summary": "Get incoming document by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Incoming Document ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/e-document-backend_internal_util.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/e-document-backend_internal_domain.IncomingDocResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/e-document-backend_internal_util.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/e-document-backend_internal_util.Response"
                        }
                    }
                }
            }
        },
        "/v1/incoming-docs/{id}/approve": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Approve or reject an incoming document (by department head or director)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "incoming-docs"
                ],
                "summary": "Approve/Reject incoming document",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Incoming Document ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Approval data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/e-document-backend_internal_domain.ApproveDocumentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/e-document-backend_internal_util.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/e-document-backend_internal_domain.IncomingDocResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/e-document-backend_internal_util.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/e-document-backend_internal_util.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/e-document-backend_internal_util.Response"
                        }
                    }
                }
            }
        },
        "/v1/incoming-docs/{id}/receive": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mark an incoming document as received (by secretary)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "incoming-docs"
                ],
                "summary": "Receive incoming document",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Incoming Document ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Receive data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/e-document-backend_internal_domain.ReceiveIncomingDocRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/e-document-backend_internal_util.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/e-document-backend_internal_domain.IncomingDocResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/e-document-backend_internal_util.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/e-document-backend_internal_util.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/e-document-backend_internal_util.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "e-document-backend_internal_domain.ApproveDocumentRequest": {
            "type": "object",
            "properties": {
                "approver_id": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "e-document-backend_internal_domain.IncomingDocResponse": {
            "type": "object",
            "properties": {
                "approver_date": {
                    "type": "string"
                },
                "approver_id": {
                    "type": "string"
                },
                "approver_name": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "dept_id": {
                    "type": "string"
                },
                "doc_id": {
                    "type": "string"
                },
                "doc_name": {
                    "type": "string"
                },
                "doc_no": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "incoming_no": {
                    "type": "string"
                },
                "received_date": {
                    "type": "string"
                },
                "receiver_id": {
                    "type": "string"
                },
                "receiver_name": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                },
                "sender_id": {
                    "type": "string"
                },
                "sender_name": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/e-document-backend_internal_domain.IncomingDocStatus"
                }
            }
        },
        "e-document-backend_internal_domain.IncomingDocStatus": {
            "type": "string",
            "enum": [
                "pending",
                "received",
                "approved",
                "rejected"
            ],
            "x-enum-varnames": [
                "IncomingStatusPending",
                "IncomingStatusReceived",
                "IncomingStatusApproved",
                "IncomingStatusRejected"
            ]
        },
...
        "e-document-backend_internal_domain.ReceiveIncomingDocRequest": {
            "type": "object",
            "properties": {
                "remark": {
                    "type": "string"
                }
            }
        },
...
        "e-document-backend_internal_util.ErrorCode": {
            "type": "string",
            "enum": [
                "UNAUTHORIZED",
                "INCORRECT_PASSWORD",
                "INVALID_CREDENTIALS",
                "TOKEN_EXPIRED",
                "INVALID_TOKEN",
                "FORBIDDEN",
                "VALIDATION_ERROR",
                "MISSING_REQUIRED_FIELD",
                "INVALID_INPUT",
                "INTERNAL_SERVER_ERROR",
                "DATABASE_ERROR",
                "CONFIG_NOT_SET",
                "USER_NOT_FOUND",
                "USER_ALREADY_EXISTS",
                "EMAIL_ALREADY_EXISTS",
                "ROLE_NOT_FOUND",
                "ROLE_ALREADY_EXISTS"
            ],
            "x-enum-varnames": [
                "UNAUTHORIZED",
                "INCORRECT_PASSWORD",
                "INVALID_CREDENTIALS",
                "TOKEN_EXPIRED",
                "INVALID_TOKEN",
                "FORBIDDEN",
                "VALIDATION_ERROR",
                "MISSING_REQUIRED_FIELD",
                "INVALID_INPUT",
                "INTERNAL_SERVER_ERROR",
                "DATABASE_ERROR",
                "CONFIG_NOT_SET",
                "USER_NOT_FOUND",
                "USER_ALREADY_EXISTS",
                "EMAIL_ALREADY_EXISTS",
                "ROLE_NOT_FOUND",
                "ROLE_ALREADY_EXISTS"
            ]
        },
...
        "e-document-backend_internal_util.PaginationInfo": {
            "type": "object",
            "properties": {
                "currentPage": {
                    "type": "integer"
                },
                "itemsPerPage": {
                    "type": "integer"
                },
                "totalItems": {
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                }
            }
        },
        "e-document-backend_internal_util.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error_code": {
                    "$ref": "#/definitions/e-document-backend_internal_util.ErrorCode"
                },
                "message": {
                    "type": "string"
                },
                "pagination": {
                    "$ref": "#/definitions/e-document-backend_internal_util.PaginationInfo"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}